Safari-609.1.11.1/LayoutTests/fast/scrolling/scroll-animator-overlay-scrollbars-clicked-expected.txt

CONSOLE MESSAGE: line 13: MainFrameView: didAddVerticalScrollbar
CONSOLE MESSAGE: line 13: MainFrameView: didAddHorizontalScrollbar
CONSOLE MESSAGE: line 13: FrameView: didAddVerticalScrollbar
CONSOLE MESSAGE: line 13: FrameView: willRemoveVerticalScrollbar
CONSOLE MESSAGE: line 16: MainFrameView: mouseEnteredContentArea
CONSOLE MESSAGE: line 16: MainFrameView: mouseMovedInContentArea
CONSOLE MESSAGE: line 18: MainFrameView: mouseEnteredVerticalScrollbar
CONSOLE MESSAGE: line 18: FrameView: mouseEnteredContentArea
CONSOLE MESSAGE: line 18: MainFrameView: mouseMovedInContentArea
CONSOLE MESSAGE: line 19: MainFrameView: mouseIsDownInVerticalScrollbar
CONSOLE MESSAGE: line 20: MainFrameView: mouseIsUpInVerticalScrollbar
CONSOLE MESSAGE: line 21: MainFrameView: mouseExitedVerticalScrollbar
CONSOLE MESSAGE: line 21: FrameView: mouseExitedContentArea
CONSOLE MESSAGE: line 21: MainFrameView: mouseMovedInContentArea
PASS successfullyParsed is true

TEST COMPLETE
Test for https://bugs.webkit.org/show_bug.cgi?id=165056.
